+++ src/server/graphql/blockchain/fetchers/messages.ts
@@ -22,6 +22,52 @@ import {
     BlockchainMessageTypeFilterEnum,
 } from "../resolvers-types-generated";
 
+
+export async function resolve_message(
+    hash: String,
+    context: QRequestContext,
+    info: GraphQLResolveInfo,
+    traceSpan: QTraceSpan,
+) {
+    const maxJoinDepth = 1;
+
+    const selectionSet = info.fieldNodes[0].selectionSet;
+    const returnExpression = config.messages.buildReturnExpression(
+        selectionSet,
+        context,
+        maxJoinDepth,
+        "doc"
+    );
+
+    // query
+    const params = new QParams();
+    const query =
+        "FOR doc IN messages " +
+        `FILTER doc._key == @${params.add(hash)} ` +
+        `RETURN ${returnExpression}`;
+    const queryResult = await context.services.data.query(
+        required(context.services.data.messages.provider),
+        {
+            text: query,
+            vars: params.values,
+            orderBy: [],
+            request: context,
+            traceSpan,
+            // TODO: shard
+        },
+    ) as BlockchainMessage[];
+
+    await config.messages.fetchJoins(
+        queryResult,
+        selectionSet,
+        context,
+        traceSpan,
+        maxJoinDepth,
+    );
+
+    return queryResult[0];
+}
+
 export async function resolve_account_messages(
     parent: BlockchainAccountQuery,
     args: BlockchainAccountQueryMessagesArgs,